chore: improve nextjs setup
