+++ examples/nextjs-with-typescript/sdk/index.tsx
@@ -1,0 +1,84 @@
+import * as React from 'react';
+import Document from 'next/document';
+import { DocumentContext } from 'next/dist/shared/lib/utils';
+import createEmotionServer from '@emotion/server/create-instance';
+import type { EmotionCache as TEmotionCache } from '@emotion/utils';
+import createCache from '@emotion/cache';
+import { CacheProvider } from '@emotion/react';
+
+export function createEmotionCache() {
+  return createCache({ key: 'css' });
+}
+
+// Client-side cache, shared for the whole session of the user in the browser.
+const clientSideEmotionCache = createEmotionCache();
+
+export type EmotionCacheAppProps = {
+  emotionCache?: TEmotionCache;
+}
+
+export function EmotionCache(props: React.PropsWithChildren<EmotionCacheAppProps>) {
+  const { emotionCache = clientSideEmotionCache } = props;
+  return (
+    <CacheProvider value={emotionCache}>
+      {props.children}
+    </CacheProvider>
+  );
+}
+
+export function withMuiSetup(opts: { getEmotionCache(): TEmotionCache; }) {
+  return async (ctx: DocumentContext) => {
+    // Resolution order
+    //
+    // On the server:
+    // 1. app.getInitialProps
+    // 2. page.getInitialProps
+    // 3. document.getInitialProps
+    // 4. app.render
+    // 5. page.render
+    // 6. document.render
+    //
+    // On the server with error:
+    // 1. document.getInitialProps
+    // 2. app.render
+    // 3. page.render
+    // 4. document.render
+    //
+    // On the client
+    // 1. app.getInitialProps
+    // 2. page.getInitialProps
+    // 3. app.render
+    // 4. page.render
+
+    const originalRenderPage = ctx.renderPage;
+
+    // You can consider sharing the same emotion cache between all the SSR requests to speed up performance.
+    // However, be aware that it can have global side effects.
+    const cache = opts.getEmotionCache();
+    const { extractCriticalToChunks } = createEmotionServer(cache);
+
+    ctx.renderPage = () =>
+      originalRenderPage({
+        enhanceApp: (App: any) => (props: any) => <App emotionCache={cache} {...props} />,
+      });
+
+    const initialProps = await Document.getInitialProps(ctx);
+    // This is important. It prevents emotion to render invalid HTML.
+    // See https://github.com/mui-org/material-ui/issues/26561#issuecomment-855286153
+    const emotionStyles = extractCriticalToChunks(initialProps.html);
+    const emotionStyleTags = emotionStyles.styles.map((style) => (
+      <style
+        data-emotion={`${style.key} ${style.ids.join(' ')}`}
+        key={style.key}
+        // eslint-disable-next-line react/no-danger
+        dangerouslySetInnerHTML={{ __html: style.css }}
+      />
+    ));
+
+    return {
+      ...initialProps,
+      // Styles fragment is rendered after the app and page rendering finish.
+      styles: [...React.Children.toArray(initialProps.styles), ...emotionStyleTags],
+    };
+  };
+}
