--- examples/nextjs-with-typescript/src/createEmotionCache.ts
@@ -1,5 +1,0 @@
-import createCache from '@emotion/cache';
-
-export default function createEmotionCache() {
-  return createCache({ key: 'css' });
-}
